requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
python-dotenv
aio_pika
pika
pytest
pytest-asyncio
flake8